Refactor settings query
