+++ src/core/react-query/settings/mutations.ts
@@ -2,10 +2,10 @@ import { useMutation } from '@tanstack/react-query';
 import jsonpatch from 'fast-json-patch';
 
 import { axios } from '@/core/axios';
-import { invalidateQueries } from '@/core/react-query/queryClient';
+import queryClient, { invalidateQueries } from '@/core/react-query/queryClient';
 
 import type { AniDBLoginRequestType, SettingsPatchRequestType } from '@/core/react-query/settings/types';
-import type { SettingsServerType } from '@/core/types/api/settings';
+import type { SettingsServerType, SettingsType } from '@/core/types/api/settings';
 
 export const useAniDBTestLoginMutation = () =>
   useMutation({
@@ -19,7 +19,8 @@ export const useCheckNetworkConnectivityMutation = () =>
 
 export const usePatchSettingsMutation = () =>
   useMutation({
-    mutationFn: ({ newSettings, oldSettings, ...params }: SettingsPatchRequestType) => {
+    mutationFn: ({ newSettings, ...params }: SettingsPatchRequestType) => {
+      const oldSettings = queryClient.getQueryData<SettingsType>(['settings'])!;
       const original: SettingsServerType = {
         ...oldSettings,
         WebUI_Settings: JSON.stringify(oldSettings.WebUI_Settings),
