+++ src/core/react-query/settings/types.ts
@@ -1,7 +1,6 @@
 import type { SettingsType } from '@/core/types/api/settings';
 
 export type SettingsPatchRequestType = {
-  oldSettings: SettingsType;
   newSettings: SettingsType;
   skipValidation?: boolean;
 };
