+++ src/pages/dashboard/panels/ContinueWatching.tsx
@@ -1,22 +1,17 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 
 import ShokoPanel from '@/components/Panels/ShokoPanel';
 import { useDashboardContinueWatchingQuery } from '@/core/react-query/dashboard/queries';
 import { useSettingsQuery } from '@/core/react-query/settings/queries';
 import EpisodeDetails from '@/pages/dashboard/components/EpisodeDetails';
-import { initialSettings } from '@/pages/settings/SettingsPage';
 
 import type { RootState } from '@/core/store';
 
 const ContinueWatching = () => {
   const layoutEditMode = useSelector((state: RootState) => state.mainpage.layoutEditMode);
 
-  const settingsQuery = useSettingsQuery();
-  const { hideR18Content } = useMemo(
-    () => settingsQuery.data?.WebUI_Settings.dashboard ?? initialSettings.WebUI_Settings.dashboard,
-    [settingsQuery],
-  );
+  const { hideR18Content } = useSettingsQuery().data.WebUI_Settings.dashboard;
 
   const items = useDashboardContinueWatchingQuery({ includeRestricted: !hideR18Content, pageSize: 20 });
