+++ src/pages/dashboard/panels/RecommendedAnime.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useSelector } from 'react-redux';
 import { mdiEyeArrowRightOutline } from '@mdi/js';
 import { Icon } from '@mdi/react';
@@ -7,7 +7,6 @@ import BackgroundImagePlaceholderDiv from '@/components/BackgroundImagePlacehold
 import ShokoPanel from '@/components/Panels/ShokoPanel';
 import { useRecommendedAnimeQuery } from '@/core/react-query/series/queries';
 import { useSettingsQuery } from '@/core/react-query/settings/queries';
-import { initialSettings } from '@/pages/settings/SettingsPage';
 
 import type { RootState } from '@/core/store';
 import type { SeriesAniDBType } from '@/core/types/api/series';
@@ -15,11 +14,7 @@ import type { SeriesAniDBType } from '@/core/types/api/series';
 const RecommendedAnime = () => {
   const layoutEditMode = useSelector((state: RootState) => state.mainpage.layoutEditMode);
 
-  const settingsQuery = useSettingsQuery();
-  const { hideR18Content } = useMemo(
-    () => settingsQuery.data?.WebUI_Settings.dashboard ?? initialSettings.WebUI_Settings.dashboard,
-    [settingsQuery],
-  );
+  const { hideR18Content } = useSettingsQuery().data.WebUI_Settings.dashboard;
 
   const items = useRecommendedAnimeQuery({
     includeRestricted: !hideR18Content,
