+++ src/pages/main/MainPage.tsx
@@ -11,15 +11,13 @@ import Header from '@/components/Layout/Header';
 import TopNav from '@/components/Layout/TopNav';
 import Events from '@/core/events';
 import { useSettingsQuery } from '@/core/react-query/settings/queries';
-import { initialSettings } from '@/pages/settings/SettingsPage';
 
 function MainPage() {
   const dispatch = useDispatch();
 
   const isSm = useMediaQuery({ minWidth: 0, maxWidth: 767 });
 
-  const settingsQuery = useSettingsQuery();
-  const { notifications, toastPosition } = settingsQuery.data?.WebUI_Settings ?? initialSettings.WebUI_Settings;
+  const { notifications, toastPosition } = useSettingsQuery().data.WebUI_Settings;
 
   const [showSmSidebar, setShowSmSidebar] = useState(false);
